+++ src/crantpy/queries/__init__.py
@@ -1,6 +1,5 @@
 import lazy_loader
 
-
 __getattr__, __dir__, __all__ = lazy_loader.attach(
     __name__,
     submodules={